fix: remove unneeded uses of ECDSA private key
Both the operator and aggregator were using the `BuildAll` helper that
required a private key due to creation of signers, only to discard the
signers and keep the reader and subscriber.
Changing it to `BuildReadClients` allows removing some of it, in the
case of the operator meaning only registration requires the ECDSA key to
be present, reducing risks of losing funds.