refactor: simplify signature timeout logic
`ProcessNewSignature` already takes a context to handle cancellation and
timeout. We need just to create the context and pass it off, and it will
return an appropriate error if the timeout expires.

This saves us from creating a goroutine and handling messages just for
that.

Also improve a little bit of the error reporting.